Add sample .envrc file (and add actual .envrc to .gitignore)
This makes some assumptions and you should edit accordingly

The first assumption is about the location of the creds repo on your
system.
The second assumption is that you have the gnu-versions of grep, etc. If
not, use `brew install coreutils` and follow the details there (or
just hardcode the vars)

Also a couple of thoughts on using the admin role in tools:

- The session duration is 1h (can we change this from our side, or it
  can only be done from the callers, i.e. gds-users side?)
- This has waaaay more permissions than our local-dev account. Might be
  worth creating a dev-role if we're going to go down this path as
  people's laptops may become compromised)