Configure template-preview for better local dev
By targetting the test image we will (when PR#752 is merged) have access
to a writeable venv for installing dependencies ad-hoc, which will make
it possible to install a local copy of utils easily, as well as use dev
tooling like ipdb/ipython/etc.

We also flag that the venv should be watched for changes to trigger hot
reloading of the flask app.