Add the sms-provider-stub as an optional container to run
Should be very rare that devs need to use this (only if they are
working on the sms stub itself) but it is nice to be able to see
that the stub works when running it locally and plugging it into
the other apps.

Hopefully I haven't missed anything else needed in here. I've
tried to copy the pattern of the other containers.

Given how rarely it will be used, I didn't end up giving it a nice
domain, you can just access the sms provider stub on
http://localhost:6300 on your browser still.