Add PORT env var for document download apps
The command for both was moved to be a script in
their repos in this pull request:

https://github.com/alphagov/notifications-local/pull/17

The `entrypoint.sh` script references the `PORT`
environment variable. When running outside of
notifications-local, `entrypoint.sh` is run by
other scripts, which set `PORT`, for example:

https://github.com/alphagov/document-download-frontend/blob/2a493399c1bd6a64a1cdbef4adea187f90da1aba/scripts/run_locally_with_docker.sh

In notifications-local, we run `entrypoint.sh`
directly so need to set it ourselves. This does
that by adding it to the env var templates for
those apps.

Note: all users who have already generated their
env var files (in `./private`) will need to add
this manually.